LPD-70458 Removed AppCreationFlow module
