--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/AppContext/actionTypes.ts
@@ -1,47 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export enum ActionTypes {
-	ADD_APP_LICENSE_PRICE = 'ADD_APP_LICENSE_PRICE',
-	DELETE_APP_LICENSE_PRICE = 'DELETE_APP_LICENSE_PRICE',
-	RESET_APP_PACKAGES = 'RESET_APP_PACKAGES',
-	SUBMIT_APP = 'SUBMIT_APP',
-	SUBMIT_APP_BUILD = 'SUBMIT_APP_BUILD',
-	SUBMIT_APP_LICENSING = 'SUBMIT_APP_LICENSING',
-	SUBMIT_APP_PROFILE = 'SUBMIT_APP_PROFILE',
-	UPDATE_APP_AREAS = 'UPDATE_APP_AREAS',
-	UPDATE_APP_BUILD = 'UPDATE_APP_BUILD',
-	UPDATE_APP_CATEGORIES = 'UPDATE_APP_CATEGORIES',
-	UPDATE_APP_DESCRIPTION = 'UPDATE_APP_DESCRIPTION',
-	UPDATE_APP_DOCUMENTATION_URL = 'UPDATE_APP_DOCUMENTATION_URL',
-	UPDATE_APP_INSTALLATION_AND_UNINSTALLATION_GUIDE_URL = 'UPDATE_APP_INSTALLATION_AND_UNINSTALLATION_GUIDE_URL',
-	UPDATE_APP_LICENSE = 'UPDATE_APP_LICENSE',
-	UPDATE_APP_LICENSE_PRICES = 'UPDATE_APP_LICENSE_PRICES',
-	UPDATE_APP_LOGO = 'UPDATE_APP_LOGO',
-	UPDATE_APP_NAME = 'UPDATE_APP_NAME',
-	UPDATE_APP_NOTES = 'UPDATE_APP_NOTES',
-	UPDATE_APP_PRICE_MODEL = 'UPDATE_APP_PRICE_MODEL',
-	UPDATE_APP_PUBLISHER_WEBSITE_URL = 'UPDATE_APP_PUBLISHER_WEBSITE_URL',
-	UPDATE_APP_SUPPORT_EMAIL = 'UPDATE_APP_SUPPORT_EMAIL',
-	UPDATE_APP_SUPPORT_PHONE = 'UPDATE_APP_SUPPORT_PHONE',
-	UPDATE_APP_SUPPORT_URL = 'UPDATE_APP_SUPPORT_URL',
-	UPDATE_APP_TAGS = 'UPDATE_APP_TAGS',
-	UPDATE_APP_TRIAL_INFO = 'UPDATE_APP_TRIAL_INFO',
-	UPDATE_APP_TYPE = 'UPDATE_APP_TYPE',
-	UPDATE_APP_USAGE_TERMS_URL = 'UPDATE_APP_USAGE_TERMS_URL',
-	UPDATE_APP_VERSION = 'UPDATE_APP_VERSION',
-	UPDATE_BUILD_PACKAGE_FILES = 'UPDATE_BUILD_PACKAGE_FILES',
-	UPDATE_CATALOG_ID = 'UPDATE_CATALOG_ID',
-	UPDATE_EULA = 'UPDATE_EULA',
-	UPDATE_EULA_CHECKBOX = 'UPDATE_EULA_CHECKBOX',
-	UPDATE_OPTION_ID = 'UPDATE_OPTION_ID',
-	UPDATE_PRODUCT_OPTION_ID = 'UPDATE_PRODUCT_OPTION_ID',
-	UPDATE_PRODUCT_OPTION_VALUES_ID = 'UPDATE_PRODUCT_OPTION_VALUES_ID',
-	UPDATE_RESOURCE_REQUIREMENTS = 'UPDATE_RESOURCE_REQUIREMENTS',
-	UPDATE_SKU_TRIAL_ID = 'UPDATE_SKU_TRIAL_ID',
-	UPDATE_SKU_VERSION_ID = 'UPDATE_SKU_VERSION_ID',
-	UPLOAD_APP_STOREFRONT_IMAGES = 'UPLOAD_APP_STOREFRONT_IMAGES',
-	UPLOAD_BUILD_PACKAGE_FILES = 'UPLOAD_BUILD_PACKAGE_FILES',
-}
