--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/AppCreationFlow.scss
@@ -1,17 +1,0 @@
-.app-creation-flow {
-	&-container {
-		display: flex;
-		flex-direction: column;
-		justify-content: space-between;
-		width: 100%;
-	}
-
-	&-body {
-		border-top: 1px solid #e7e7ed;
-		display: flex;
-		flex-direction: row;
-		gap: 5rem;
-		justify-content: center;
-		padding-top: 3.5rem;
-	}
-}
