--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ChoosePricingModelPage/ChoosePricingModelPage.scss
@@ -1,11 +1,0 @@
-.choose-pricing-model-page {
-	&-container {
-		width: 37.5rem;
-	}
-
-	&-radio-container {
-		display: flex;
-		flex-direction: column;
-		gap: 1.5rem;
-	}
-}
