--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/CreateNewAppPage/CreateNewAppPage.scss
@@ -1,9 +1,0 @@
-.create-new-app {
-	&-container {
-		width: 37.5rem;
-	}
-
-	&-header {
-		margin-bottom: 2.5rem;
-	}
-}
