--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/DefineAppProfilePage/DefineAppProfilePage.scss
@@ -1,77 +1,0 @@
-.profile-page {
-	&-container {
-		width: 600px;
-
-		.form-control {
-			background-color: #ffff;
-			border: 1px solid #b1b2b9;
-			border-radius: 8px;
-			min-height: 48px;
-
-			.label {
-				background-color: #e6ebf5;
-				border: none;
-				color: #1c3667;
-				font-size: 13px;
-				padding: 8px;
-			}
-		}
-	}
-
-	&-body-container {
-		.upload-logo-container {
-			justify-content: flex-start;
-		}
-	}
-}
-
-input[type='file'] {
-	display: none;
-}
-
-.upload-logo {
-	&-container {
-		align-items: center;
-		display: flex;
-		gap: 1rem;
-		justify-content: center;
-		margin-bottom: 1.5rem;
-	}
-
-	&-icon {
-		border-radius: 4px;
-		max-height: 4.5rem;
-		object-fit: contain;
-		width: 4.5rem;
-	}
-
-	&-upload-label {
-		border-radius: 6px;
-		transition: background-color 0.3s;
-
-		&:hover {
-			background-color: #0a58e7;
-		}
-	}
-
-	&-delete-button {
-		border: 1.2px solid #2b3a4b;
-		border-radius: 0.1875rem;
-		transition: background-color 0.3s;
-
-		&-text {
-			color: #2b3a4b;
-			font-size: 0.875rem;
-			font-weight: 600;
-			line-height: 1rem;
-
-			transition: color 0.3s;
-		}
-
-		&:hover {
-			background-color: #2b3a4b;
-			border: 1.5px solid #2b3a4b;
-			color: #fff;
-		}
-	}
-}
