--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/InformLicensingTermsPage/InformLicensingTermsPage.scss
@@ -1,29 +1,0 @@
-.informing-licensing-terms-page {
-	&-container {
-		width: 37.5rem;
-
-		.icon-button {
-			background: #fbfcfe;
-			border: 1px solid #e7efff;
-			border-radius: 10px;
-			color: #0b5fff;
-			font-size: 14px;
-			font-weight: 600;
-			line-height: 16px;
-		}
-	}
-
-	&-app-license-container {
-		display: flex;
-		flex-direction: column;
-		gap: 1.5rem;
-		margin-bottom: 2.5rem;
-	}
-
-	&-day-trial-container {
-		align-items: center;
-		display: flex;
-		gap: 1.5rem;
-		justify-content: flex-start;
-	}
-}
