--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/InformLicensingTermsPage/components/CurrencySelector/CurrencySelector.scss
@@ -1,17 +1,0 @@
-.currency-selector {
-	span {
-		color: var(--neutral-10, #282934);
-		font-size: 14px;
-		font-weight: 600;
-	}
-
-	svg {
-		color: #999aa3;
-	}
-
-	&-icon {
-		flex-shrink: 0;
-		height: 20px;
-		width: 20px;
-	}
-}
