--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/InformLicensingTermsPage/components/LicensePriceCard/LicensePriceCard.scss
@@ -1,16 +1,0 @@
-.license-card {
-	&-container {
-		border: 1px solid #e2e2e4;
-		border-radius: 10px;
-	}
-
-	&-input {
-		border: 1px solid var(--neutral-07, #6c6c75);
-		border-radius: 8px !important;
-
-		&.disabled {
-			background: var(--neutral-01, #f7f7f8);
-			border: 1px solid var(--neutral-03, #c9c9cf);
-		}
-	}
-}
