--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/InformLicensingTermsPage/utils/formatCurrency.ts
@@ -1,24 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {Liferay} from '../../../../../../../liferay/liferay';
-
-const formatCurrency = (
-	value: number,
-	currencyKey?: string,
-	precision?: String
-) => {
-	return new Intl.NumberFormat(Liferay.ThemeDisplay.getBCP47LanguageId(), {
-		currency: currencyKey ? currencyKey : 'USD',
-
-		// @ts-ignore
-
-		roundingPriority: precision || 'morePrecision',
-
-		style: 'currency',
-	}).format(value);
-};
-
-export default formatCurrency;
