--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ProvideAppBuildPage/ResourceRequirements/index.scss
@@ -1,13 +1,0 @@
-.resource-requirements-content {
-	&-input {
-		height: 48px;
-		width: 288px;
-	}
-
-	label::after {
-		color: #da1414;
-		content: '*';
-		font-size: 16px;
-		font-weight: bold;
-	}
-}
