--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ProvideAppSupportAndHelpPage/ProvideAppSupportAndHelpPage.scss
@@ -1,5 +1,0 @@
-.provide-app-support-and-help-page {
-	&-container {
-		width: 37.5rem;
-	}
-}
