--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ProvideVersionDetailsPage/ProvideVersionDetailsPage.scss
@@ -1,5 +1,0 @@
-.provide-version-details-page {
-	&-container {
-		width: 37.5rem;
-	}
-}
