--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ReviewAndSubmitAppPage/CardSection.scss
@@ -1,55 +1,0 @@
-.card-section {
-	&-container {
-		border: 1px solid #e2e2e4;
-		border-radius: 0.625rem;
-		width: 37.5rem;
-	}
-
-	&-body-section {
-		border-bottom: 1px solid #e2e2e4;
-		padding-bottom: 33px;
-		padding-top: 33px;
-
-		&:last-child {
-			border-bottom: none;
-		}
-
-		&-header {
-			align-items: center;
-			display: flex;
-			justify-content: space-between;
-
-			&-actions {
-				display: flex;
-				gap: 12px;
-
-				.edit-button {
-					color: #0b5fff;
-					font-size: 14px;
-					font-weight: 600;
-					line-height: 16px;
-				}
-			}
-
-			.field-base-localized-field {
-				button {
-					color: #2b3a4b;
-				}
-
-				.tooltip-base-container {
-					margin-bottom: 0;
-				}
-
-				.arrow-down-icon {
-					filter: invert(20%) sepia(24%) saturate(334%)
-						hue-rotate(196deg) brightness(92%) contrast(82%);
-				}
-			}
-
-			&-title {
-				font-size: 18px;
-				font-weight: 600;
-			}
-		}
-	}
-}
