--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ReviewAndSubmitAppPage/CardSectionsBody.scss
@@ -1,170 +1,0 @@
-.card-section-body-section {
-	&-file {
-		align-items: center;
-		display: flex;
-		padding-top: 20px;
-
-		&-container {
-			align-items: center;
-			background-color: #e7efff;
-			border-radius: 4px;
-			display: flex;
-			height: 72px;
-			justify-content: center;
-			width: 72px;
-
-			&-icon {
-				height: 24px;
-				width: 24px;
-				filter: invert(35%) sepia(86%) saturate(1411%)
-					hue-rotate(206deg) brightness(102%) contrast(101%);
-			}
-		}
-
-		&-icon {
-			filter: invert(26%) sepia(83%) saturate(5507%) hue-rotate(220deg)
-				brightness(104%) contrast(97%);
-			height: 16px;
-			margin-left: 16px;
-			width: 16px;
-		}
-
-		&-name {
-			color: #282934;
-			font-size: 16px;
-			font-weight: 600;
-			line-height: 24px;
-			margin-left: 4px;
-		}
-	}
-
-	&-files {
-		display: flex;
-		gap: 16px;
-		padding-top: 20px;
-		width: 100%;
-
-		&-container {
-			align-items: center;
-			background-color: #fff;
-			border-radius: 4px;
-			display: flex;
-			height: 128px;
-			justify-content: center;
-			width: 220px;
-
-			.preview-image {
-				background-position: 50% 50%;
-				background-size: cover;
-				border-radius: 0.625rem;
-				max-height: 100%;
-				max-width: 100%;
-				object-fit: contain;
-			}
-		}
-
-		&-data {
-			display: flex;
-			flex-direction: column;
-			width: 314px;
-
-			&-icon {
-				filter: invert(26%) sepia(83%) saturate(5507%)
-					hue-rotate(220deg) brightness(104%) contrast(97%);
-				height: 16px;
-				margin-bottom: 4px;
-				width: 16px;
-			}
-
-			&-name {
-				color: #282934;
-				font-size: 16px;
-				font-weight: 600;
-				line-height: 24px;
-				margin-bottom: 4px;
-			}
-
-			&-description {
-				word-break: break-all;
-			}
-		}
-
-		&-info {
-			font-size: 13px;
-			line-height: 16px;
-			margin-top: 16px;
-		}
-	}
-
-	&-paragraph {
-		margin-bottom: 0;
-		margin-top: 22px;
-		word-break: break-word;
-		word-wrap: break-word;
-	}
-
-	&-requirements {
-		gap: 24px;
-	}
-
-	&-tags {
-		display: flex;
-		flex-wrap: wrap;
-		gap: 8px;
-		margin-top: 22px;
-
-		.tag {
-			white-space: nowarp;
-		}
-	}
-
-	&-version {
-		display: flex;
-		gap: 16px;
-		padding-top: 20px;
-		width: 100%;
-
-		&-container {
-			align-items: center;
-			background: #e6ebf5;
-			border-radius: 4px;
-			display: flex;
-			height: 80px;
-			justify-content: center;
-			width: 106px;
-
-			&-icon {
-				align-items: center;
-				align-self: center;
-				background-color: #2e5aac;
-				border-radius: 4px;
-				color: #fff;
-				display: flex;
-				font-size: 13px;
-				font-weight: 400;
-				height: 24px;
-				justify-content: center;
-				line-height: 16px;
-				width: 42px;
-			}
-		}
-
-		&-data {
-			display: flex;
-			flex-direction: column;
-			width: 430px;
-
-			&-name {
-				color: #282934;
-				font-size: 16px;
-				font-weight: 600;
-				line-height: 24px;
-				margin-bottom: 4px;
-			}
-
-			&-description {
-				word-break: break-all;
-			}
-		}
-	}
-}
