--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ReviewAndSubmitAppPage/ReviewAndSubmitAppPage.scss
@@ -1,104 +1,0 @@
-.review-and-submit-app-page {
-	&-container {
-		width: 37.5rem;
-	}
-
-	&-card {
-		&-container {
-			border: 1px solid #e2e2e4;
-			border-radius: 0.625rem;
-			width: 37.5rem;
-		}
-
-		&-header {
-			align-items: center;
-			display: flex;
-			justify-content: space-between;
-			padding: 1.5rem 1.5rem 0;
-
-			&-button {
-				background: none;
-				border: 1px solid #2b3a4b;
-				border-radius: 0.375rem;
-				font-size: 0.875rem;
-				font-weight: 600;
-				height: 2rem;
-				line-height: 1rem;
-				transition: background-color 0.3s;
-				transition: color 0.3s;
-				width: 6.625rem;
-
-				&:hover {
-					background-color: #2b3a4b;
-					color: #fff;
-				}
-			}
-
-			&-icon {
-				background-position: 50% 50%;
-				background-repeat: 'no-repeat';
-				background-size: cover;
-				border-radius: 10px;
-				max-height: 6.5rem;
-				object-fit: scale-down;
-				width: 6.5rem;
-			}
-
-			&-title {
-				display: flex;
-				flex-direction: column;
-
-				&-text {
-					color: #282934;
-					font-size: 24px;
-					font-weight: 700;
-					line-height: 28px;
-					word-break: break-word;
-					word-wrap: break-word;
-				}
-
-				&-version {
-					color: #54555f;
-					font-size: 14px;
-					line-height: 16px;
-					margin-top: 6px;
-				}
-			}
-
-			&-left-content {
-				display: flex;
-				gap: 1rem;
-			}
-		}
-
-		&-body {
-			display: flex;
-			flex-direction: column;
-			padding: 0.5rem 1.5rem;
-		}
-	}
-
-	&-header {
-		margin-bottom: 2.5rem;
-	}
-
-	&-agreement {
-		display: flex;
-
-		&-highlight {
-			font-weight: 600;
-		}
-	}
-
-	&-info-footer {
-		color: #54555f;
-		font-size: 0.813rem;
-		font-weight: 400;
-		line-height: 1rem;
-
-		&-container {
-			margin-left: 0.313rem;
-			margin-top: 1.563rem;
-		}
-	}
-}
