--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/StorefrontPage/CustomizeAppStorefrontPage.scss
@@ -1,35 +1,0 @@
-.storefront-page {
-	&-container {
-		width: 37.5rem;
-	}
-
-	&-info {
-		&-container {
-			align-items: center;
-			display: flex;
-			justify-content: space-between;
-			margin-bottom: 1.5rem;
-		}
-
-		&-text {
-			color: #282934;
-			font-size: 1rem;
-			font-weight: 600;
-			line-height: 1.5rem;
-		}
-
-		&-button {
-			background-color: transparent;
-			border: none;
-			border-radius: 0.375rem;
-			color: #0b5fff;
-			height: 2rem;
-			transition: background-color 0.3s;
-			width: 5.813rem;
-
-			&:hover {
-				background-color: #edf3fe;
-			}
-		}
-	}
-}
