+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/NewAppFlow/constants/productTypes.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {ProductType} from '../../../../../../../enums/Product';
-import i18n from '../../../../../../../i18n';
+import {ProductType} from '../../../../../enums/Product';
+import i18n from '../../../../../i18n';
 
 type ProductTypeOption = {
 	description: string;
