+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/NewAppFlow/pages/Build.tsx
@@ -19,7 +19,7 @@ import {
 	ProductWorkflowStatusCode,
 } from '../../../../../enums/Product';
 import i18n from '../../../../../i18n';
-import {ProductTypeOptions} from '../../Apps/AppCreationFlow/ProvideAppBuildPage/constants/productTypes';
+import {ProductTypeOptions} from '../constants/productTypes';
 import CloudResourceRequirements from '../components/CloudResourceRequirements';
 import {NewAppPackageVersionModal} from '../components/NewAppPackagesModal';
 import NewAppUploadAppPackagesComponent from '../components/NewAppUploadPackage';
