+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/NewAppFlow/pages/Storefront.tsx
@@ -18,7 +18,7 @@ import {ProductTags} from '../../../../../enums/Product';
 import i18n from '../../../../../i18n';
 import {swapElements} from '../../../../../utils/array';
 import {getRandomID} from '../../../../../utils/string';
-import {ACCEPT_FILE_TYPES} from '../../Apps/AppCreationFlow/StorefrontPage/CustomizeAppStorefrontPage';
+import {ACCEPT_FILE_TYPES} from '../../../../../enums/File';
 import {MAX_IMAGE_QUANTITY, MAX_SIZE_5MBS} from '../constants';
 
 const Storefront = () => {
