+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Solutions/SolutionForm/pages/Header.tsx
@@ -22,7 +22,7 @@ import {
 import i18n from '../../../../../../i18n';
 import {swapElements} from '../../../../../../utils/array';
 import {getRandomID} from '../../../../../../utils/string';
-import {ACCEPT_FILE_TYPES} from '../../../Apps/AppCreationFlow/StorefrontPage/CustomizeAppStorefrontPage';
+import {ACCEPT_FILE_TYPES} from '../../../../../../enums/File';
 import {MAX_IMAGE_QUANTITY, MAX_SIZE_5MBS} from '../../constants';
 
 enum ContentMediaType {
